+++ eabm/eabm_package/python/single_element_compression_nominal.py
@@ -1,6 +1,6 @@
 simulation_variables = {
-    '@width@': 1.0,
-    '@height@': 1.0,
-    '@global_seed@': 1.0,
-    '@displacement@': -1.0
+    'width': 1.0,
+    'height': 1.0,
+    'global_seed': 1.0,
+    'displacement': -1.0
 }