+++ docs/changelog.rst
@@ -8,6 +8,11 @@ Changelog
 0.2.3 (unreleased)
 *******************
 
+Breaking changes
+================
+- Change the parameter study data key from 'values' to 'samples' to avoid name clash with the 'values' method and
+  attribute of dictionaries and datasets. (:issue:`234`, :merge:`229`). By `Kyle Brindley`_.
+
 *******************
 0.2.2 (2022-07-28)
 *******************