+++ docs/changelog.rst
@@ -22,7 +22,7 @@ Breaking changes
 New Features
 ============
 - Create a ``waves`` command line utility with a version argument and a subparser for opening the packaged HTML 
-  documentation in the system default web browser (:issue:`172`, :merge:`233`). By `Thomas Robert`_.
+  documentation in the system default web browser (:issue:`172`, :merge:`233`). By `Thomas Roberts`_.
 
 Bug fixes
 =========