+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Documentation
   :merge:`237`). By `Kyle Brindley`_.
 - Update the tutorial discussions about the simulation variables dictionary usage (:issue:`243`, :merge:`241`). By `Kyle
   Brindley`_.
+- Standardize the WAVES-EABM parameter set module names (:issue:`245`, :merge:`242`). By `Kyle Brindley`_.
 
 Internal Changes
 ================
@@ -32,6 +33,7 @@ Internal Changes
   file import to CAE. (:issue:`240`, :merge:`238`). By `Kyle Brindley`_.
 - In WAVES-EABM, use parameter name keys without substitution syntax and perform substitution syntax key string changes
   only when necessary for parameter substitution (:issue:`243`, :merge:`241`). By `Kyle Brindley`_.
+- Standardize the WAVES-EABM parameter set module names (:issue:`245`, :merge:`242`). By `Kyle Brindley`_.
 
 ******************
 0.3.1 (2022-08-02)