+++ docs/changelog.rst
@@ -42,7 +42,7 @@ Internal Changes
   only when necessary for parameter substitution (:issue:`243`, :merge:`241`). By `Kyle Brindley`_.
 - Standardize the WAVES-EABM parameter set module names (:issue:`245`, :merge:`242`). By `Kyle Brindley`_.
 - Standardize code snippet markers in tutorial configuration files to reduce diff clutter in the documentation
-  (:issue:`247`, :merge:`244`, :merge:`245`). By `Kyle Brindley`_.
+  (:issue:`247`, :merge:`245`). By `Kyle Brindley`_.
 
 ******************
 0.3.1 (2022-08-02)